+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.3.0
 
+- fixed materials not being packaged
 - do not include duplicates in imported asset output and total
 - sort imported assets by filesize
 - build button